requirements.txt
bcrypt==3.2.0
mysql-connector-python==8.0.28
nltk==3.6.5
